1.1.0
1. Добавлена возможность работы с несколькими столбцами заметок.
2. Доработан функционал окна сообщений.
3. Добвлен скрипт установщика с автоматическим добавлением приложения в автозапуск.
4. Приложение переписано под паттерн MVVM.
5. Добавлена иконка приложения.

+++ Notes/Models/SettingsManager.cs
@@ -32,7 +32,10 @@ namespace Notes.Models
     {
         public static void Save<T>(string FilePath, T Data)
         {
-            File.WriteAllText(FilePath, string.Empty);
+            if (File.Exists(FilePath))
+            {
+                File.WriteAllText(FilePath, string.Empty);
+            }            
 
             JsonSerializerOptions JsonOptions = new JsonSerializerOptions()
             {
